training/requirements.txt
datasets
pandas
pyarrow
pillow
matplotlib
tqdm
pyyaml
transformers
torch
seqeval
scikit-learn
